backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
pydantic==2.5.3
pydantic-settings==2.1.0
anthropic==0.18.1
httpx==0.26.0
python-dotenv==1.0.0
PyGithub==2.1.1
beautifulsoup4==4.12.3
lxml==5.1.0
html2text==2024.2.26
requests==2.31.0
gitpython==3.1.41
asyncio==3.4.3
aiofiles==23.2.1
python-multipart==0.0.6

